pyproject.toml
[build-system]
requires = ["setuptools>=45", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "kdp-formatter"
dynamic = ["version"]
description = "A tool for formatting books according to Amazon KDP standards"
readme = "README.md"
requires-python = ">=3.8"
license = { text = "MIT" }
dependencies = [
    "Pillow",
    "opencv-python",
    "pypandoc",
    "python-docx",
    "ebooklib",
    "reportlab",
    "beautifulsoup4",
    "PyPDF2",
    "pdf2image",
    "svglib",
    "cairosvg",
]

[tool.setuptools]
packages = ["kdp_formatter"]